backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
gTTS==2.4.0
pypinyin==0.50.0
pykakasi==2.2.1
hangul-romanize==0.1.0
aiofiles==23.2.1
pandas==2.1.3
deep-translator==1.11.4
google-generativeai==0.8.3
APScheduler==3.10.4